+++ standard-plugin/src/main/kotlin/dragons/plugins/standard/vulkan/model/matrices/DragonMatrices.kt
@@ -4,7 +4,7 @@ import dragons.plugins.standard.vulkan.model.generator.dragon.DragonWingProperti
 import org.joml.Matrix4f
 
 fun createDragonWingMatrices(baseMatrix: Matrix4f, shoulderMatrix: Matrix4f, props: DragonWingProperties): Array<Matrix4f> {
-    val directionBasedElbow = baseMatrix.translate(0f, 0f, props.wingTopLength.meters * 2, Matrix4f())
+    val directionBasedElbow = baseMatrix.translate(0f, 0f, props.wingTopLength.meters.toFloat() * 2, Matrix4f())
 
     fun mix(mixer: Float, matrix0: Matrix4f, matrix1: Matrix4f): Matrix4f {
         return matrix0.lerp(matrix1, mixer, Matrix4f())
